refactor(4c4): extract classic standings layout into share/layouts
Moves the 400-line JSX from /api/share/grid/[gameId]/route.tsx into a
pure classicStandingsLayout() function with overflow capping (20 alive +
10 eliminated). The route is now a thin handler that calls into the
layout, making it reusable for the future /standings route (Task 5).
